fix: allow for legacy definition import timing
